pyproject.toml
[tool.poetry]
name = "fine-grained-emotions-bert"
version = "0.1.0"
description = ""
authors = ["jeongwon412 <jeongwon412@gmail.com>"]
readme = "README.md"
packages = [{include = "fine_grained_emotions_bert"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.13.0"
pandas = "^1.5.1"
numpy = "^1.23.5"
seaborn = "^0.12.1"
transformers = "^4.24.0"
jupyterlab = "^3.5.0"
ipywidgets = "^8.0.2"
pytest = "^7.2.0"
pyyaml = "^6.0"
black = "^22.10.0"

[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
